+++ hmri_create_b1map.m
@@ -97,9 +97,14 @@ end
 %   otherwise copyfile does not find the files!!
 
 if ~isempty(P_trans)
+    if ~isfield(jobsubj,'b1_suffix')
+        b1_suffix = '';
+    else
+        b1_suffix = jobsubj.b1_suffix;
+    end
     P_trans = spm_file(P_trans,'number','');
-    P_trans_copy{1} = fullfile(jobsubj.path.b1respath, spm_file(P_trans(1,:), 'filename'));
-    P_trans_copy{2} = fullfile(jobsubj.path.b1respath, spm_file(P_trans(2,:), 'filename'));
+    P_trans_copy{1} = fullfile(jobsubj.path.b1respath, spm_file(spm_file(P_trans(1,:), 'filename'),'suffix',b1_suffix));
+    P_trans_copy{2} = fullfile(jobsubj.path.b1respath, spm_file(spm_file(P_trans(2,:), 'filename'),'suffix',b1_suffix));
     copyfile(deblank(P_trans(1,:)), P_trans_copy{1});
     try copyfile([spm_str_manip(P_trans(1,:),'r') '.json'],[spm_str_manip(P_trans_copy{1},'r') '.json']); end %#ok<*TRYNC>
     copyfile(deblank(P_trans(2,:)), P_trans_copy{2});